Update database schema for WuwaMainStats and WuwaSubStats
Modified the `Name` property of `WuwaMainStats` to have a maximum length of 100 characters and changed its column type to `character varying(100)`.

Introduced a new entity `WuwaSubStats` with auto-generated `Id` and seeded initial data for both `WuwaMainStats` and `WuwaSubStats`.

Updated the migration file `20250519125957_WuwaStatSeedling.cs` to define the changes in the `Up` method and provide a rollback in the `Down` method.

The designer file was also updated to reflect the new model structure for Entity Framework Core.