Remove data seeding for Games and GameStat tables
This commit removes data seeding for the `Games`, `GameStat`, and `StatTypes` tables in the `AppDbContextModelSnapshot.cs` file. A new migration class named `DeleteDataSeeding` has been added, which includes methods to delete the specified data and restore it if necessary. The `Designer.cs` file for the migration has also been updated accordingly.