Add migration class FixGameStatsName
Introduce a new migration class `FixGameStatsName` in the `API.Migrations` namespace. This class includes empty `Up` and `Down` methods. A designer file is also created to define the target model for the migration, outlining the structure of entities such as `Game`, `GameArtifactName`, `GameStat`, `StatType`, and `WuwaMainStat`, along with their properties, keys, and relationships.